requirements.txt
Django==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.7.0
asgiref==3.8.1
sqlparse==0.5.3
PyJWT==2.10.1
